@@ -1,4 +1,7 @@
+"""Clients and Drivers classes."""
+
 __all__ = ["AsyncClient", "AsyncDriver", "Client", "Driver", "Websocket"]
-from .driver import Driver, AsyncDriver
-from .client import Client, AsyncClient
+
+from .client import AsyncClient, Client
+from .driver import AsyncDriver, Driver
 from .websocket import Websocket
