@@ -1,45 +1,44 @@
-import asyncio
-import logging
-import warnings
+"""Driver classes."""
 
-from .client import AsyncClient, Client
+from asyncio import AbstractEventLoop, get_event_loop
+from logging import DEBUG, INFO, Logger, getLogger
+from typing import Any, Dict
+
+from requests import Response
+
+from .client import AsyncClient, Client, ClientType
+from .endpoints import *
 from .websocket import Websocket
-from .endpoints.brand import Brand
-from .endpoints.channels import Channels
-from .endpoints.cluster import Cluster
-from .endpoints.commands import Commands
-from .endpoints.compliance import Compliance
-from .endpoints.files import Files
-from .endpoints.ldap import LDAP
-from .endpoints.oauth import OAuth
-from .endpoints.posts import Posts
-from .endpoints.preferences import Preferences
-from .endpoints.reactions import Reactions
-from .endpoints.saml import SAML
-from .endpoints.system import System
-from .endpoints.teams import Teams
-from .endpoints.users import Users
-from .endpoints.webhooks import Webhooks
-from .endpoints.elasticsearch import Elasticsearch
-from .endpoints.emoji import Emoji
-from .endpoints.data_retention import DataRetention
-from .endpoints.roles import Roles
-from .endpoints.status import Status
-from .endpoints.bots import Bots
-from .endpoints.opengraph import Opengraph
-from .endpoints.integration_actions import IntegrationActions
 
-log = logging.getLogger("mattermostdriver.api")
-log.setLevel(logging.INFO)
+log: Logger = getLogger("mattermostdriver.api")
+log.setLevel(INFO)
 
 
 class BaseDriver:
-    """
-    Contains the client, api and provides you with functions for
-    login, logout and initializing a websocket connection.
+    """Base for creating driver classes.
+
+    Contains the client, api and provides you with functions for login, logout
+    and initializing a websocket connection.
+
+    Attributes
+    ----------
+    options : dict
+        The options for driver and client.
+    driver : dict
+        The options for driver.
+    client : AsyncClient or Client
+        The underlying client object.
+    websocket : Websocket, default=None
+        The websocket to listen to Mattermost events.
+
+    Methods
+    -------
+    disconnect()
+        Disconnect the driver from the server.
+
     """
 
-    default_options = {
+    default_options: Dict[str, Any] = {
         "scheme": "https",
         "url": "localhost",
         "port": 8065,
@@ -60,282 +59,370 @@ class BaseDriver:
         "proxy": None,
     }
     """
-	Required options
-		- url
+    Required options
+        - url
 
-	Either
-		- login_id
-		- password
+    Either
+        - login_id
+        - password
 
-	Or
-		- token (https://docs.mattermost.com/developer/personal-access-tokens.html)
+    Or
+        - token (https://docs.mattermost.com/developer/personal-access-tokens.html)
 
-	Optional
-		- scheme ('https')
-		- port (8065)
-		- verify (True)
-		- timeout (30)
-		- request_timeout (None)
-		- mfa_token (None)
-		- auth (None)
-		- debug (False)
+    Optional
+        - scheme ('https')
+        - port (8065)
+        - verify (True)
+        - timeout (30)
+        - request_timeout (None)
+        - mfa_token (None)
+        - auth (None)
+        - debug (False)
 
-	Should not be changed
-		- basepath ('/api/v4') - unlikely this would do any good
-	"""
+    Should not be changed
+        - basepath ('/api/v4') - unlikely this would do any good
+    """
 
-    def __init__(self, options, client_cls):
+    def __init__(
+        self, options: Dict[str, Any], client_cls: ClientType
+    ) -> None:
+        """Initialize driver.
+
+        Parameters
+        ----------
+        options : dict
+            The options for driver and client.
+        client_cls : AsyncClient or Client
+            The underlying client class.
+
         """
-        :param options: A dict with the values from `default_options`
-        :type options: dict
-        """
-        self.options = self.default_options.copy()
-        if options is not None:
+        self.options: Dict[str, Any] = self.default_options.copy()
+
+        if options:
             self.options.update(options)
-        self.driver = self.options
-        if self.options["debug"]:
-            log.setLevel(logging.DEBUG)
+
+        self.driver: Dict[str, Any] = self.options
+
+        if self.options.get("debug"):
+            log.setLevel(DEBUG)
             log.warning(
                 "Careful!!\nSetting debug to True, will reveal your password in the log output if you do driver.login()!\nThis is NOT for production!"
             )
-        self.client = client_cls(self.options)
-        self.websocket = None
 
-    def disconnect(self):
-        """Disconnects the driver from the server, stopping the websocket event loop."""
-        self.websocket.disconnect()
+        self.client: ClientType = client_cls(self.options)
+        self.websocket: Websocket | None = None
+
+    # ############################################################ Properties #
 
     @property
-    def users(self):
-        """
-        Api endpoint for users
+    def users(self) -> Users:
+        """Get Api endpoint for users.
 
-        :return: Instance of :class:`~endpoints.users.Users`
+        Returns
+        -------
+        endpoints.users.Users
+
         """
         return Users(self.client)
 
     @property
-    def teams(self):
-        """
-        Api endpoint for teams
+    def teams(self) -> Teams:
+        """Get Api endpoint for teams.
 
-        :return: Instance of :class:`~endpoints.teams.Teams`
+        Returns
+        -------
+        endpoints.teams.Teams
+
         """
         return Teams(self.client)
 
     @property
-    def channels(self):
-        """
-        Api endpoint for channels
+    def channels(self) -> Channels:
+        """Get Api endpoint for channels.
 
-        :return: Instance of :class:`~endpoints.channels.Channels`
+        Returns
+        -------
+        endpoints.channels.Channels
+
         """
         return Channels(self.client)
 
     @property
-    def posts(self):
-        """
-        Api endpoint for posts
+    def posts(self) -> Posts:
+        """Get Api endpoint for posts.
 
-        :return: Instance of :class:`~endpoints.posts.Posts`
+        Returns
+        -------
+        endpoints.posts.Posts
+
         """
         return Posts(self.client)
 
     @property
-    def files(self):
-        """
-        Api endpoint for files
+    def files(self) -> Files:
+        """Get Api endpoint for files.
 
-        :return: Instance of :class:`~endpoints.files.Files`
+        Returns
+        -------
+        endpoints.files.Files
+
         """
         return Files(self.client)
 
     @property
-    def preferences(self):
-        """
-        Api endpoint for preferences
+    def preferences(self) -> Preferences:
+        """Get Api endpoint for preferences.
 
-        :return: Instance of :class:`~endpoints.preferences.Preferences`
+        Returns
+        -------
+        endpoints.preferences.Preferences
+
         """
         return Preferences(self.client)
 
     @property
-    def emoji(self):
-        """
-        Api endpoint for emoji
+    def emoji(self) -> Emoji:
+        """Get Api endpoint for emoji.
 
-        :return: Instance of :class:`~endpoints.emoji.Emoji`
+        Returns
+        -------
+        endpoints.emoji.Emoji
+
         """
         return Emoji(self.client)
 
     @property
-    def reactions(self):
-        """
-        Api endpoint for posts' reactions
+    def reactions(self) -> Reactions:
+        """Get Api endpoint for posts' reactions.
 
-        :return: Instance of :class:`~endpoints.reactions.Reactions`
+        Returns
+        -------
+        endpoints.reactions.Reactions
+
         """
         return Reactions(self.client)
 
     @property
-    def system(self):
-        """
-        Api endpoint for system
+    def system(self) -> System:
+        """Get Api endpoint for system.
 
-        :return: Instance of :class:`~endpoints.system.System`
+        Returns
+        -------
+        endpoints.system.System
+
         """
         return System(self.client)
 
     @property
-    def webhooks(self):
-        """
-        Api endpoint for webhooks
+    def webhooks(self) -> Webhooks:
+        """Get Api endpoint for webhooks.
 
-        :return: Instance of :class:`~endpoints.webhooks.Webhooks`
+        Returns
+        -------
+        endpoints.webhooks.Webhooks
+
         """
         return Webhooks(self.client)
 
     @property
-    def compliance(self):
-        """
-        Api endpoint for compliance
+    def compliance(self) -> Compliance:
+        """Get Api endpoint for compliance.
 
-        :return: Instance of :class:`~endpoints.compliance.Compliance`
+        Returns
+        -------
+        endpoints.compliance.Compliance
+
         """
         return Compliance(self.client)
 
     @property
-    def cluster(self):
-        """
-        Api endpoint for cluster
+    def cluster(self) -> Cluster:
+        """Get Api endpoint for cluster.
 
-        :return: Instance of :class:`~endpoints.cluster.Cluster`
+        Returns
+        -------
+        endpoints.cluster.Cluster
+
         """
         return Cluster(self.client)
 
     @property
-    def brand(self):
-        """
-        Api endpoint for brand
+    def brand(self) -> Brand:
+        """Get Api endpoint for brand.
 
-        :return: Instance of :class:`~endpoints.brand.Brand`
+        Returns
+        -------
+        endpoints.brand.Brand
+
         """
         return Brand(self.client)
 
     @property
-    def oauth(self):
-        """
-        Api endpoint for oauth
+    def oauth(self) -> OAuth:
+        """Get Api endpoint for oauth.
 
-        :return: Instance of :class:`~endpoints.oauth.OAuth`
+        Returns
+        -------
+        endpoints.oauth.OAuth
+
         """
         return OAuth(self.client)
 
     @property
-    def saml(self):
-        """
-        Api endpoint for saml
+    def saml(self) -> SAML:
+        """Get Api endpoint for SAML.
 
-        :return: Instance of :class:`~endpoints.saml.SAML`
+        Returns
+        -------
+        endpoints.saml.SAML
+
         """
         return SAML(self.client)
 
     @property
-    def ldap(self):
-        """
-        Api endpoint for ldap
+    def ldap(self) -> LDAP:
+        """Get Api endpoint for LDAP.
 
-        :return: Instance of :class:`~endpoints.ldap.LDAP`
+        Returns
+        -------
+        endpoints.ldap.LDAP
+
         """
         return LDAP(self.client)
 
     @property
-    def elasticsearch(self):
-        """
-        Api endpoint for elasticsearch
+    def elasticsearch(self) -> Elasticsearch:
+        """Get Api endpoint for ElasticSearch.
 
-        :return: Instance of :class:`~endpoints.elasticsearch.Elasticsearch`
+        Returns
+        -------
+        endpoints.elasticsearch.Elasticsearch
+
         """
         return Elasticsearch(self.client)
 
     @property
-    def data_retention(self):
-        """
-        Api endpoint for data_retention
+    def data_retention(self) -> DataRetention:
+        """Get Api endpoint for data_retention.
 
-        :return: Instance of :class:`~endpoints.data_retention.DataRetention`
+        Returns
+        -------
+        endpoints.data_retention.DataRetention
+
         """
         return DataRetention(self.client)
 
     @property
-    def status(self):
-        """
-        Api endpoint for status
+    def status(self) -> Status:
+        """Get Api endpoint for status.
 
-        :return: Instance of :class:`~endpoints.status.Status`
+        Returns
+        -------
+        endpoints.status.Status
+
         """
         return Status(self.client)
 
     @property
-    def commands(self):
-        """
-        Api endpoint for commands
+    def commands(self) -> Commands:
+        """Get Api endpoint for commands.
 
-        :return: Instance of :class:`~endpoints.commands.Commands`
+        Returns
+        -------
+        endpoints.commands.Commands
+
         """
         return Commands(self.client)
 
     @property
-    def roles(self):
-        """
-        Api endpoint for roles
+    def roles(self) -> Roles:
+        """Get Api endpoint for roles.
 
-        :return: Instance of :class:`~endpoints.roles.Roles`
+        Returns
+        -------
+        endpoints.roles.Roles
+
         """
         return Roles(self.client)
 
     @property
-    def opengraph(self):
-        """
-        Api endpoint for opengraph
+    def opengraph(self) -> Opengraph:
+        """Get Api endpoint for opengraph.
 
-        :return: Instance of :class:`~endpoints.opengraph.Opengraph`
+        Returns
+        -------
+        endpoints.opengraph.Opengraph
+
         """
         return Opengraph(self.client)
 
     @property
-    def integration_actions(self):
-        """
-        Api endpoint for integration actions
+    def integration_actions(self) -> IntegrationActions:
+        """Get Api endpoint for integration actions.
 
-        :return: Instance of :class:`~endpoints.integration_actions.IntegrationActions`
+        Returns
+        -------
+        endpoints.integration_actions
+
         """
         return IntegrationActions(self.client)
 
     @property
-    def bots(self):
-        """
-        Api endpoint for bots
+    def bots(self) -> Bots:
+        """Get Api endpoint for bots.
 
-        :return: Instance of :class:`~endpoints.bots.Bots`
+        Returns
+        -------
+        endpoints.bots.Bots
+
         """
         return Bots(self.client)
 
+    # ############################################################### Methods #
+
+    def disconnect(self) -> None:
+        """Disconnect the driver from the server.
+
+        It stops the websocket event loop.
+
+        """
+        if self.websocket:
+            self.websocket.disconnect()
+
 
 class Driver(BaseDriver):
-    def __init__(self, options=None, client_cls=Client):
+    """Class defining a synchronous Mattermost driver.
+
+    Methods
+    -------
+    init_websocket(event_handler, websocket_cls)
+        Initialize the websocket connection to the Mattermost server.
+    login()
+        Log the user in.
+    logout()
+        Log the user out.
+
+    """
+
+    def __init__(
+        self,
+        options: Dict[str, Any] | None = None,
+        client_cls: ClientType = Client,
+    ) -> None:
         super().__init__(options, client_cls)
 
     def __enter__(self):
         self.client.__enter__()
+
         return self
 
-    def __exit__(self, *exc_info):
+    def __exit__(self, *exc_info) -> Any:
         return self.client.__exit__(*exc_info)
 
-    def init_websocket(self, event_handler, websocket_cls=Websocket):
-        """
-        Will initialize the websocket connection to the mattermost server.
+    def init_websocket(
+        self, event_handler: Any, websocket_cls: Websocket = Websocket
+    ) -> AbstractEventLoop:
+        """Initialize the websocket connection to the Mattermost server.
 
         This should be run after login(), because the websocket needs to make
         an authentification.
@@ -350,32 +437,51 @@ class Driver(BaseDriver):
                 async def my_event_handler(message):
                         print(message)
 
+        Parameters
+        ----------
+        event_handler : Function(message)
+            The function to handle the websocket events. Takes one argument.
+        websocket_cls : websocket.Websocket, default=websocket.Websocket
+            The Websocket class.
 
-        :param event_handler: The function to handle the websocket events. Takes one argument.
-        :type event_handler: Function(message)
-        :return: The event loop
+        Returns
+        -------
+        asyncio.AbstractEventLoop
+            The event loop.
+
         """
         self.websocket = websocket_cls(self.options, self.client.token)
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(self.websocket.connect(event_handler))
+        loop: AbstractEventLoop = get_event_loop()
+
+        try:
+            loop.run_until_complete(self.websocket.connect(event_handler))
+
+        except RuntimeError as err:
+            log.error(err)
+
         return loop
 
-    def login(self):
-        """
-        Logs the user in.
+    def login(self) -> Any | Response:
+        """Log the user in.
 
         The log in information is saved in the client
-                - userid
-                - username
-                - cookies
+        - userid
+        - username
+        - cookies
 
-        :return: The raw response from the request
+        Returns
+        -------
+        Any or requests.Response
+            The reponse in JSON format or the raw response if couldn't be
+            converted to JSON.
+
         """
-        if self.options["token"]:
+        if self.options.get("token"):
             self.client.token = self.options["token"]
-            result = self.users.get_user("me")
+            result: Any = self.users.get_user("me")
+
         else:
-            response = self.users.login_user(
+            response: Response = self.users.login_user(
                 {
                     "login_id": self.options["login_id"],
                     "password": self.options["password"],
@@ -385,50 +491,78 @@ class Driver(BaseDriver):
             if response.status_code == 200:
                 self.client.token = response.headers["Token"]
                 self.client.cookies = response.cookies
+
             try:
                 result = response.json()
+
             except ValueError:
-                log.debug("Could not convert response to json, returning raw response")
+                log.debug(
+                    "Could not convert response to json, returning raw response"
+                )
                 result = response
 
-        log.debug(result)
-
-        if "id" in result:
+        if result.get("id"):
             self.client.userid = result["id"]
-        if "username" in result:
+
+        if result.get("username"):
             self.client.username = result["username"]
 
         return result
 
-    def logout(self):
-        """
-        Log the user out.
+    def logout(self) -> Any:
+        """Log the user out.
 
-        :return: The JSON response from the server
+        Returns
+        -------
+        Any
+            The reponse in JSON format.
+
         """
-        result = self.users.logout_user()
+        result: Any = self.users.logout_user()
+
         self.client.token = ""
         self.client.userid = ""
         self.client.username = ""
         self.client.cookies = None
+
         return result
 
 
 class AsyncDriver(BaseDriver):
-    def __init__(self, options=None, client_cls=AsyncClient):
+    """Class defining an asynchronous Mattermost driver.
+
+    Methods
+    -------
+    init_websocket(event_handler, websocket_cls)
+        Initialize the websocket connection to the Mattermost server.
+    login()
+        Log the user in.
+    logout()
+        Log the user out.
+
+    """
+
+    def __init__(
+        self,
+        options: Dict[str, Any] | None = None,
+        client_cls: ClientType = AsyncClient,
+    ) -> None:
         super().__init__(options, client_cls)
 
     async def __aenter__(self):
         await self.client.__aenter__()
+
         return self
 
-    async def __aexit__(self, *exc_info):
+    async def __aexit__(self, *exc_info) -> Any:
         return await self.client.__aexit__(*exc_info)
 
-    def init_websocket(self, event_handler, websocket_cls=Websocket):
-        """
-        Will initialize the websocket connection to the mattermost server.
-        unlike the Driver.init_websocket, this one assumes you are async aware
+    def init_websocket(
+        self, event_handler: Any, websocket_cls: Websocket = Websocket
+    ) -> Any:
+        """Initialize the websocket connection to the Mattermost server.
+
+        Unlike the Driver.init_websocket, this one assumes you are async aware
         and returns a coroutine that can be awaited.  It will not return
         until shutdown() is called.
 
@@ -445,63 +579,86 @@ class AsyncDriver(BaseDriver):
                 async def my_event_handler(message):
                         print(message)
 
+        Parameters
+        ----------
+        event_handler : Function(message)
+            The function to handle the websocket events. Takes one argument.
+        websocket_cls : websocket.Websocket, default=websocket.Websocket
+            The Websocket class.
 
-        :param event_handler: The function to handle the websocket events. Takes one argument.
-        :type event_handler: Function(message)
-        :return: coroutine
+        Returns
+        -------
+        Coroutine
+
         """
         self.websocket = websocket_cls(self.options, self.client.token)
+
         return self.websocket.connect(event_handler)
 
-    async def login(self):
-        """
-        Logs the user in.
+    async def login(self) -> Any | Response:
+        """Log the user in.
 
         The log in information is saved in the client
-                - userid
-                - username
-                - cookies
+        - userid
+        - username
+        - cookies
 
-        :return: The raw response from the request
+        Returns
+        -------
+        Any or requests.Response
+            The reponse in JSON format or the raw response if couldn't be
+            converted to JSON.
+
         """
-        if self.options["token"]:
+        if self.options.get("token"):
             self.client.token = self.options["token"]
-            result = await self.users.get_user("me")
+            result: Any = await self.users.get_user("me")
+
         else:
-            response = await self.users.login_user(
+            response: Any = await self.users.login_user(
                 {
                     "login_id": self.options["login_id"],
                     "password": self.options["password"],
                     "token": self.options["mfa_token"],
                 }
             )
+
             if response.status_code == 200:
                 self.client.token = response.headers["Token"]
                 self.client.cookies = response.cookies
+
             try:
                 result = response.json()
+
             except ValueError:
-                log.debug("Could not convert response to json, returning raw response")
+                log.debug(
+                    "Could not convert response to json, returning raw response"
+                )
                 result = response
 
-        log.debug(result)
-
-        if "id" in result:
+        if result("id"):
             self.client.userid = result["id"]
-        if "username" in result:
+
+        if result.get("username"):
             self.client.username = result["username"]
 
         return result
 
-    async def logout(self):
-        """
-        Log the user out.
+    async def logout(self) -> Any:
+        """Log the user out.
 
-        :return: The JSON response from the server
+        Returns
+        -------
+        Any or requests.Response
+            The reponse in JSON format or the raw response if couldn't be
+            converted to JSON.
+
         """
-        result = await self.users.logout_user()
+        result: Any = await self.users.logout_user()
+
         self.client.token = ""
         self.client.userid = ""
         self.client.username = ""
         self.client.cookies = None
+
         return result
