@@ -1,2 +1,4 @@
-full_version = "7.3.2"
-short_version = ".".join(full_version.split(".", 2)[:2])
+"""Project's version."""
+
+full_version: str = "7.3.2"
+short_version: str = ".".join(full_version.split(".", 2)[:2])
